% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{common_filters}
\alias{common_filters}
\title{helpers}
\usage{
common_filters(additional_filters = NULL)
}
\arguments{
\item{additional_filters}{UI elements for tab-specific filters. These will appear in the
same row as the common filters.}
}
\value{
The return value, if any, from executing the function.

A `fluidRow` containing shared `selectInput` elements (year and reviewer)
  and any additional filters passed to the function.
}
\description{
A fct function
}
\examples{
# Use this function in a Shiny UI:
common_filters(
  shiny::selectInput("specific_filter", "Specific Filter", choices = c("A", "B"))
)

}
