% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{custom_query_duckdb}
\alias{custom_query_duckdb}
\title{Execute a Custom Query on a MotherDuck DuckDB Database}
\usage{
custom_query_duckdb(query)
}
\arguments{
\item{query}{A character string containing the SQL query to execute.}
}
\value{
A data frame containing the results of the SQL query.
}
\description{
This function connects to a MotherDuck DuckDB database, executes a custom SQL query,
and returns the queried data as a data frame. The connection is automatically closed after the query.
}
\examples{
# Example usage:
query <- "SELECT * FROM awards WHERE year > 2020"
results <- custom_query_duckdb(query)

}
